+++ src/main/java/com/autoever/recall/auth/config/SecurityConfigImpl.java
@@ -21,7 +21,7 @@ import java.util.List;
 
 @Configuration
 @EnableWebSecurity
-@Profile("local")
+@Profile("!local")
 @RequiredArgsConstructor
 public class SecurityConfigImpl implements SecurityConfig {
     private final AuthenticationEntryPoint authenticationEntryPoint;
