+++ src/main/java/com/autoever/recall/auth/service/AuthService.java
@@ -3,7 +3,5 @@ package com.autoever.recall.auth.service;
 import org.springframework.security.core.Authentication;
 
 public interface AuthService {
-    Long getSessionUserId();
-
     Authentication login(String email, String password);
 }
