+++ src/main/java/com/autoever/recall/auth/service/SecuritySessionService.java
@@ -1,0 +1,5 @@
+package com.autoever.recall.auth.service;
+
+public interface SecuritySessionService {
+    Long getSessionUserId();
+}
