+++ src/main/java/com/autoever/recall/userschool/service/UserSchoolServiceImpl.java
@@ -1,6 +1,6 @@
 package com.autoever.recall.userschool.service;
 
-import com.autoever.recall.auth.service.AuthService;
+import com.autoever.recall.auth.service.SecuritySessionService;
 import com.autoever.recall.school.domain.SchoolType;
 import com.autoever.recall.school.service.SchoolService;
 import com.autoever.recall.userschool.domain.UserSchool;
@@ -17,12 +17,12 @@ import java.util.List;
 @Transactional(readOnly = true)
 public class UserSchoolServiceImpl implements UserSchoolService {
     private final UserSchoolRepository userSchoolRepository;
-    private final AuthService authService;
+    private final SecuritySessionService securitySessionService;
     private final SchoolService schoolService;
 
     @Override
     public UserSchool getMySchool(SchoolType type) {
-        Long userId = authService.getSessionUserId();
+        Long userId = securitySessionService.getSessionUserId();
         return userSchoolRepository.findByUserIdAndSchoolType(userId, type)
                                    .orElseThrow(() -> new UserSchoolNotFoundException(userId, type));
     }
